requirements.txt
numpy>=1.26,<2
torch>=2.1
matplotlib>=3.8
openenv-core==0.2.3
fastapi>=0.110
uvicorn[standard]>=0.30
transformers>=4.40
trl>=0.18
datasets>=2.18
accelerate>=0.30
peft>=0.11
sentencepiece>=0.2
huggingface_hub>=0.23
